usbpcap: ignore IRP Information and Unknown packets.
This solves Issue #3, though it does not add a test because I have been
unable to reproduce this in an isolated environment that would give us
a safe test data file to use.

While we could probably expose at least the IRP Information as its own
structure, usbmon-tools is aimed at dealing with the higher-level USB
packets, rather than the lower level traces that IRP Information would
be useful with.